+++ apps/scoutgame/lib/builders/getTopBuilders.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@charmverse/core/prisma-client';
-import { currentSeason } from '@packages/scoutgame/utils';
+import { currentSeason } from '@packages/scoutgame/dates';
 import { isTruthy } from '@root/lib/utils/types';
 
 import { seasonQualifiedBuilderWhere } from './queries';
