+++ apps/scoutgame/lib/scouts/getScoutStats.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@charmverse/core/prisma-client';
-import { currentSeason } from '@packages/scoutgame/utils';
+import { currentSeason } from '@packages/scoutgame/dates';
 
 export async function getScoutStats(scoutId: string) {
   const scout = await prisma.scout.findUniqueOrThrow({
