+++ apps/scoutgame/lib/scouts/getScoutedBuilders.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@charmverse/core/prisma-client';
-import { currentSeason, getCurrentWeek } from '@packages/scoutgame/utils';
+import { currentSeason, getCurrentWeek } from '@packages/scoutgame/dates';
 
 import { BasicUserInfoSelect } from 'lib/users/queries';
