+++ apps/scoutgame/lib/users/getUserStats.ts
@@ -1,6 +1,6 @@
 import type { UserAllTimeStats, UserSeasonStats } from '@charmverse/core/prisma-client';
 import { prisma } from '@charmverse/core/prisma-client';
-import { currentSeason } from '@packages/scoutgame/utils';
+import { currentSeason } from '@packages/scoutgame/dates';
 
 export type UserStats = {
   githubLogin?: string;
