+++ apps/scoutgamecron/src/tasks/processGemsPayout/processScoutPointsPayout.ts
@@ -1,6 +1,6 @@
 import { prisma } from '@charmverse/core/prisma-client';
 import { calculateEarnableScoutPointsForRank } from '@packages/scoutgame/calculatePoints';
-import { currentSeason } from '@packages/scoutgame/utils';
+import { currentSeason } from '@packages/scoutgame/dates';
 
 export async function processScoutPointsPayout({
   builderId,
