+++ packages/scoutgame/src/builderNfts/getAllNftOwners.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@charmverse/core/prisma-client';
 import { arrayUtils } from '@charmverse/core/utilities';
 
-import { currentSeason } from '../utils';
+import { currentSeason } from '../dates';
 
 export async function getAllNftOwners({ builderId }: { builderId: string }): Promise<string[]> {
   const builderNft = await prisma.builderNft.findUniqueOrThrow({
