+++ packages/scoutgame/src/builderNfts/refreshBuilderNftPrice.ts
@@ -2,7 +2,8 @@ import { InvalidInputError } from '@charmverse/core/errors';
 import { prisma } from '@charmverse/core/prisma-client';
 import { stringUtils } from '@charmverse/core/utilities';
 import { getPublicClient } from '@packages/onchain/getPublicClient';
-import { currentSeason } from '@packages/scoutgame/utils';
+
+import { currentSeason } from '../dates';
 
 import { builderContractAddress, builderNftChain } from './constants';
 import { ContractApiClient } from './nftContractApiClient';
