+++ packages/scoutgame/src/dates.ts
