+++ packages/scoutgame/src/refreshUserStats.ts
@@ -2,7 +2,7 @@ import type { Prisma, UserSeasonStats, UserWeeklyStats } from '@charmverse/core/
 import { prisma } from '@charmverse/core/prisma-client';
 import { arrayUtils } from '@charmverse/core/utilities';
 
-import { currentSeason, getCurrentWeek } from './utils';
+import { currentSeason, getCurrentWeek } from './dates';
 
 export async function refreshUserStats({
   userId
