+++ packages/scoutgame/src/testing/database.ts
@@ -2,7 +2,7 @@ import type { BuilderEvent, BuilderEventType, GithubRepo } from '@charmverse/cor
 import { GithubEventType, prisma } from '@charmverse/core/prisma-client';
 import { v4 as uuid } from 'uuid';
 
-import { currentSeason, getCurrentWeek } from '../utils';
+import { currentSeason, getCurrentWeek } from '../dates';
 
 import { randomLargeInt } from './generators';
